Move modes to individual files
