+++ src/lib.rs
@@ -71,364 +71,34 @@ pub mod num;
 pub mod ui;
 pub mod storage;
 
-mod engine;
+mod mode;
 
-pub use engine::Outcome;
 pub use storage::Explorer;
 
-use engine::{Order, Failure, Mode};
+use mode::{Pane, Flag, Operation, Output, Processor};
 use pancurses::Input;
-use rec::ChCls::{Any, Digit, End, Not, Sign, Whitespace};
-use rec::{lazy_some, opt, some, tkn, var, Element, Pattern};
 use std::borrow::Borrow;
 use std::cell::RefCell;
-use std::cmp::{self, Ordering};
+use std::cmp::Ordering;
 use std::collections::HashMap;
-use std::error;
 use std::fmt::{self, Debug, Display, Formatter};
-use std::mem::{self, Discriminant};
-use std::iter;
-use std::ops::{Add, AddAssign, Shr, ShrAssign, Sub};
-use std::path::{Path, PathBuf};
+use std::ops::{Add, AddAssign};
 use std::rc::Rc;
 use try_from::{TryFrom, TryFromIntError};
-use ui::{
-    Address, Change, Color, Edit, Index, IndexType, Length, Region, UserInterface, BACKSPACE,
-    ENTER, ESC,
-};
-use std::hash::{Hash, Hasher};
+use ui::UserInterface;
 
-/// An [`IndexType`] with a value of `-1`.
-const NEGATIVE_ONE: IndexType = -1;
+type Mrc<T> = Rc<RefCell<T>>;
 
 /// The paper application.
 #[derive(Debug)]
 pub struct Paper {
     /// User interface of the application.
     ui: Rc<dyn UserInterface>,
-    /// The current [`Mark`]s where [`Edit`]s may be made.
-    marks: Vec<Mark>,
-    /// The current [`Mode`] of the application.
-    mode: Mode,
-    mode_handlers: HashMap<Mode, Rc<RefCell<dyn ModeHandler>>>,
-}
-
-trait ModeHandler: Debug {
-    fn enter(&mut self, order: Option<Order>) -> Outcome<Vec<Edit>>;
-    fn process(&mut self, input: char) -> Outcome<Operation>;
-}
-
-#[derive(Clone, Debug)]
-struct DisplayModeHandler {
-    ui: Rc<dyn UserInterface>,
+    pane: Rc<RefCell<Pane>>,
     explorer: Rc<RefCell<dyn Explorer>>,
-    view: Rc<RefCell<View>>,
-}
-
-enum Operation {
-    ChangeMode(Mode, Option<Order>),
-    EditUi(Vec<Edit>),
-    Noop,
-}
-
-impl DisplayModeHandler {
-    fn new(ui: &Rc<dyn UserInterface>, explorer: Rc<RefCell<dyn Explorer>>) -> Self {
-        Self {
-            ui: Rc::clone(ui),
-            explorer: explorer,
-            view: Rc::new(RefCell::new(View::default())),
-        }
-    }
-
-    /// Returns the height used for scrolling.
-    #[allow(clippy::integer_arithmetic)] // okay to divide usize by 4
-    fn scroll_height(&self) -> Result<usize, TryFromIntError> {
-        self.ui.grid_height().map(|height| height / 4)
-    }
-
-    fn view_edits(&self) -> Outcome<Vec<Edit>> {
-        Ok(self.view.borrow_mut().redraw_edits().take(self.ui.grid_height()?).collect())
-    }
-}
-
-impl ModeHandler for DisplayModeHandler {
-    fn enter(&mut self, order: Option<Order>) -> Outcome<Vec<Edit>> {
-        match order {
-            Some(Order::SetView(path)) => {
-                let absolute_path = if path.is_absolute() {
-                    path.to_path_buf()
-                } else {
-                    let mut new_path = std::env::current_dir()?;
-                    new_path.push(path);
-                    new_path
-                };
-
-                self.view.borrow_mut().change(
-                    Rc::clone(&self.explorer),
-                    absolute_path,
-                )?;
-                //self.noises.clear();
-
-                //for line in 1..=self.view.borrow_mut().line_count {
-                //    if let Some(noise) = LineNumber::new(line).map(Section::line) {
-                //        self.noises.push(noise);
-                //    }
-                //}
-            }
-            Some(Order::Save) => {
-                let view = self.view.borrow_mut();
-                self.explorer.borrow_mut().write(&view.path, &view.data)?;
-            }
-            _ => (),
-        }
-
-        self.view_edits()
-    }
-
-    fn process(&mut self, input: char) -> Outcome<Operation> {
-        return match input {
-            '.' => {
-                Ok(Operation::ChangeMode(Mode::Command, None))
-            }
-            '#' | '/' => {
-                let view = Rc::make_mut(&mut self.view).clone();
-                Ok(Operation::ChangeMode(Mode::Filter, Some(Order::StartFilter(input, view.into_inner(), self.ui.grid_height()?))))
-            }
-            'j' => {
-                let movement = IndexType::try_from(self.scroll_height()?)?;
-
-                if self.view.borrow_mut().scroll(movement) {
-                    return self.view_edits().map(Operation::EditUi);
-                }
-
-                Ok(Operation::Noop)
-            }
-            'k' => {
-                let mut movement = IndexType::try_from(self.scroll_height()?)?;
-                movement = movement
-                    .checked_neg()
-                    .ok_or(Failure::Conversion(TryFromIntError::Overflow))?;
-
-                if self.view.borrow_mut().scroll(movement) {
-                    return self.view_edits().map(Operation::EditUi);
-                }
-
-                Ok(Operation::Noop)
-            }
-            _ => Ok(Operation::Noop),
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct CommandModeHandler {
-    command: String,
-    command_pattern: Pattern,
-}
-
-impl CommandModeHandler {
-    fn new() -> Self {
-        CommandModeHandler {
-            command: String::new(),
-            command_pattern: Pattern::define(
-                tkn!(lazy_some(Any) => "command")
-                    + (End | (some(Whitespace) + tkn!(var(Any) => "args"))),
-            ),
-        }
-    }
-
-    fn command_edits(&self) -> Vec<Edit> {
-        vec![Edit::new(Region::with_row(0).unwrap(), Change::Row(self.command.clone()))]
-    }
-}
-
-impl ModeHandler for CommandModeHandler {
-    fn enter(&mut self, _order: Option<Order>) -> Outcome<Vec<Edit>> {
-        self.command.clear();
-        Ok(self.command_edits())
-    }
-
-    fn process(&mut self, input: char) -> Outcome<Operation> {
-        return match input {
-            ENTER => {
-                let mut command = Order::Draw;
-                let command_tokens = self.command_pattern.tokenize(&self.command);
-
-                match command_tokens.get("command") {
-                    Some("see") => {
-                        if let Some(path) = command_tokens.get("args") {
-                            command = Order::SetView(PathBuf::from(path));
-                        }
-                    }
-                    Some("put") => {
-                        command = Order::Save;
-                    }
-                    Some("end") => {
-                        return Err(Failure::Quit);
-                    }
-                    Some(_) | None => {}
-                }
-
-                Ok(Operation::ChangeMode(Mode::Display, Some(command)))
-            }
-            ESC => {
-                Ok(Operation::ChangeMode(Mode::Display, None))
-            }
-            _ => {
-                if input == BACKSPACE {
-                    if self.command.pop().is_none() {
-                        return Ok(Operation::EditUi(vec![Edit::new(Region::default(), Change::Flash)]));
-                    }
-                } else {
-                    self.command.push(input);
-                }
-
-                Ok(Operation::EditUi(self.command_edits()))
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-struct FilterModeHandler {
-    filter: String,
-    noises: Vec<Section>,
-    signals: Vec<Section>,
-    first_feature_pattern: Pattern,
-    filters: PaperFilters,
-    view: View,
-    ui_height: usize,
-}
-
-impl FilterModeHandler {
-    fn new() -> Self {
-        Self {
-            filter: String::new(),
-            noises: Vec::new(),
-            signals: Vec::new(),
-            first_feature_pattern: Pattern::define(tkn!(var(Not("&")) => "feature") + opt("&&")),
-            filters: PaperFilters::default(),
-            view: View::default(),
-            ui_height: 0,
-        }
-    }
-
-    fn filter_edits(&self) -> Vec<Edit> {
-        let mut edits = vec![Edit::new(Region::with_row(0).unwrap(), Change::Row(self.filter.clone()))];
-
-        for row in 0..self.ui_height {
-            edits.push(self.format_region_edit(Region::with_row(row).unwrap(), Color::Default));
-        }
-
-        for noise in &self.noises {
-            if let Some(edit) = self.format_section_edit(noise, Color::Blue) {
-                edits.push(edit);
-            }
-        }
-
-        for signal in &self.signals {
-            if let Some(edit) = self.format_section_edit(signal, Color::Red) {
-                edits.push(edit);
-            }
-        }
-
-        edits
-    }
-
-    fn format_section_edit(&self, section: &Section, color: Color) -> Option<Edit> {
-        // It is okay for region_at() to return None; this just means section is not displayed.
-        self.view.region_at(section).map(|region| self.format_region_edit(region, color))
-    }
-
-    fn format_region_edit(&self, region: Region, color: Color) -> Edit {
-        Edit::new(region, Change::Format(color))
-    }
-}
-
-impl ModeHandler for FilterModeHandler {
-    fn enter(&mut self, order: Option<Order>) -> Outcome<Vec<Edit>> {
-        self.filter.clear();
-        
-        if let Some(Order::StartFilter(c, view, height)) = order {
-            self.filter.push(c);
-            self.view = view;
-            self.ui_height = height;
-        }
-
-        
-        Ok(vec![Edit::new(
-            Region::with_row(0)?,
-            Change::Row(self.filter.clone()),
-        )])
-    }
-
-    fn process(&mut self, input: char) -> Outcome<Operation> {
-        return match input {
-            ENTER => {
-                Ok(Operation::ChangeMode(Mode::Action, None))
-            }
-            ESC => {
-                Ok(Operation::ChangeMode(Mode::Display, None))
-            }
-            _ => {
-                if input == BACKSPACE {
-                    if self.filter.pop().is_none() {
-                        return Ok(Operation::EditUi(vec![Edit::new(Region::default(), Change::Flash)]));
-                    }
-                } else if input == '\t' {
-                    self.filter.push_str("&&");
-                    self.noises = self.signals.clone();
-                } else {
-                    self.filter.push(input);
-                }
-
-                if let Some(last_feature) = self
-                    .first_feature_pattern
-                    .tokenize_iter(&self.filter)
-                    .last()
-                    .and_then(|tokens| tokens.get("feature"))
-                {
-                    self.signals = self.noises.clone();
-
-                    if let Some(id) = last_feature.chars().nth(0) {
-                        for filter in self.filters.iter() {
-                            if id == filter.id() {
-                                filter.extract(last_feature, &mut self.signals, &self.view)?;
-                                break;
-                            }
-                        }
-                    }
-                }
-
-                Ok(Operation::EditUi(self.filter_edits()))
-            }
-        };
-    }
-}
-
-#[derive(Debug)]
-struct ActionModeHandler {
-}
-
-impl ModeHandler for ActionModeHandler {
-    fn enter(&mut self, order: Option<Order>) -> Outcome<Vec<Edit>> {
-        Ok(vec![])
-    }
-
-    fn process(&mut self, input: char) -> Outcome<Operation> {
-        return match input {
-            ESC => Ok(Operation::ChangeMode(Mode::Display, None)),
-            'i' => {
-                //self.set_marks(Edge::Start);
-                Ok(Operation::ChangeMode(Mode::Edit, None))
-            }
-            'I' => {
-                //self.set_marks(Edge::End);
-                Ok(Operation::ChangeMode(Mode::Edit, None))
-            }
-            _ => Ok(Operation::Noop),
-        };
-    }
+    /// The current [`mode::Name`] of the application.
+    mode: mode::Name,
+    mode_handlers: HashMap<mode::Name, Rc<RefCell<dyn Processor>>>,
 }
 
 impl Paper {
@@ -436,27 +106,35 @@ impl Paper {
     #[inline]
     pub fn new(ui: Rc<dyn UserInterface>, explorer: Rc<RefCell<dyn Explorer>>) -> Self {
         explorer.borrow_mut().start().unwrap();
-        let display_mode_handler: Rc<RefCell<dyn ModeHandler>> = Rc::new(RefCell::new(DisplayModeHandler::new(&ui, explorer)));
-        let command_mode_handler: Rc<RefCell<dyn ModeHandler>> = Rc::new(RefCell::new(CommandModeHandler::new()));
+        let pane = Rc::new(RefCell::new(Pane::new(ui.grid_height().unwrap())));
+        let display_mode_handler: Rc<RefCell<dyn Processor>> = Rc::new(RefCell::new(mode::DisplayProcessor::new(&pane, &explorer)));
+        let command_mode_handler: Rc<RefCell<dyn Processor>> = Rc::new(RefCell::new(mode::CommandProcessor::new()));
+        let filter_mode_handler: Rc<RefCell<dyn Processor>> = Rc::new(RefCell::new(mode::FilterProcessor::new(&pane)));
+        let action_mode_handler: Rc<RefCell<dyn Processor>> = Rc::new(RefCell::new(mode::ActionProcessor::new(&pane)));
+        let edit_mode_handler: Rc<RefCell<dyn Processor>> = Rc::new(RefCell::new(mode::EditProcessor::new(&pane)));
 
         Self {
             ui,
-            mode: Mode::default(),
+            pane,
+            explorer,
+            mode: mode::Name::default(),
             mode_handlers: [
-                (Mode::Display, display_mode_handler),
-                (Mode::Command, command_mode_handler),
+                (mode::Name::Display, display_mode_handler),
+                (mode::Name::Command, command_mode_handler),
+                (mode::Name::Filter, filter_mode_handler),
+                (mode::Name::Action, action_mode_handler),
+                (mode::Name::Edit, edit_mode_handler),
             ].iter().cloned().collect(),
-            marks: Vec::default(),
         }
     }
 
     /// Runs the application.
     #[inline]
-    pub fn run(&mut self) -> Outcome<()> {
+    pub fn run(&mut self) -> Output<()> {
         self.ui.init()?;
 
         loop {
-            if let Err(Failure::Quit) = self.step() {
+            if let Err(Flag::Quit) = self.step() {
                 break;
             }
         }
@@ -466,856 +144,28 @@ impl Paper {
     }
 
     /// Processes 1 input from the user.
-    pub fn step(&mut self) -> Outcome<()> {
-        let mut input = '\0';
-        
-        if let Some(Input::Character(c)) = self.ui.receive_input() {
-            input = c;
-        }
+    pub fn step(&mut self) -> Output<()> {
+        if let Some(Input::Character(input)) = self.ui.receive_input() {
+            let operation = self.current_processor_mut().borrow_mut().decode(input)?;
 
-        let operation = self.mode_handlers.get_mut(&self.mode).unwrap().borrow_mut().process(input)?;
-
-        match operation {
-            Operation::ChangeMode(new_mode, order) => {
-                for edit in self.mode_handlers.get_mut(&new_mode).unwrap().borrow_mut().enter(order)? {
-                    self.ui.apply(edit)?;
+            let edits = match operation {
+                Operation::EnterMode(new_mode, initiation) => {
+                    self.mode = new_mode;
+                    self.current_processor_mut().borrow_mut().enter(initiation)?
                 }
+                Operation::EditUi(ui_edits) => ui_edits,
+                _ => vec![],
+            };
 
-                self.mode = new_mode;
-            }
-            Operation::EditUi(edits) => {
-                for edit in edits {
-                    self.ui.apply(edit)?;
-                }
-            }
-            _ => (),
-        }
-
-        //match (self.mode, self.ui.receive_input()) {
-        //    (Mode::Edit, Some(Input::Character(c))) => {
-        //        if c == ESC {
-        //            self.display_view()?;
-        //            self.change_mode(Mode::Display);
-        //        } else {
-        //            self.update_view(c)?;
-        //        }
-        //    }
-        //    (_, _) => {}
-        //}
-
-        Ok(())
-    }
-
-    /// Displays the view on the user interface.
-    fn display_view(&self) -> Outcome<()> {
-        //for edit in self.view.borrow_mut().redraw_edits().take(self.ui.grid_height()?) {
-        //    self.ui.apply(edit)?;
-        //}
-
-        Ok(())
-    }
-
-    /// Sets the [`Mark`]s of the application to be at the given [`Edge`] of all signals.
-    fn set_marks(&mut self, edge: Edge) {
-        //self.marks.clear();
-
-        //for signal in &self.signals {
-        //    let mut place = signal.start;
-
-        //    if edge == Edge::End {
-        //        place.column += Index::try_from(signal.length)
-        //            .unwrap_or_else(|_| self.view.borrow_mut().line_length(signal.start).unwrap_or_default())
-        //    }
-
-        //    let pointer = Pointer(
-        //        self.view.borrow_mut()
-        //            .line_indices()
-        //            .nth(place.line.row())
-        //            .and_then(|index_value| Index::try_from(index_value).ok()),
-        //    ) + place.column;
-        //    self.marks.push(Mark { place, pointer });
-        //}
-    }
-
-    /// Formats a region to a given [`Color`].
-    fn format_region(&self, region: Region, color: Color) -> ui::Outcome {
-        self.ui.apply(Edit::new(region, Change::Format(color)))
-    }
-
-    /// Adds a char to all [`Mark`]s and updates the view.
-    fn update_view(&mut self, c: char) -> Outcome<()> {
-        //let mut adjustment = Adjustment::default();
-
-        //for mark in &mut self.marks {
-        //    if let Some(new_adjustment) = Adjustment::create(c, mark.place, &self.view.borrow_mut()) {
-        //        adjustment += new_adjustment;
-
-        //        if adjustment.change != Change::Clear {
-        //            if let Some(region) = self.view.borrow_mut().region_at(&mark.place) {
-        //                self.ui
-        //                    .apply(Edit::new(region, adjustment.change.clone()))?;
-        //            }
-        //        }
-
-        //        mark.adjust(&adjustment);
-        //        self.view.borrow_mut().add(mark, c)?;
-        //    }
-        //}
-
-        //if adjustment.change == Change::Clear {
-        //    self.view.borrow_mut().clean();
-        //    self.display_view()?;
-        //}
-
-        Ok(())
-    }
-
-    /// Changes the mode of the application.
-    fn change_mode(&mut self, mode: Mode) {
-        self.mode = mode;
-    }
-}
-
-/// Signifies all of the [`Filters`] used by the application.
-#[derive(Debug, Default)]
-struct PaperFilters {
-    /// The [`Filter`] that matches lines.
-    line: LineFilter,
-    /// The [`Filter`] that matches patterns.
-    pattern: PatternFilter,
-}
-
-impl PaperFilters {
-    /// Returns the [`Iterator`] of [`Filters`].
-    fn iter(&self) -> PaperFiltersIter<'_> {
-        PaperFiltersIter {
-            index: 0,
-            filters: self,
-        }
-    }
-}
-
-/// Signifies an [`Iterator`] through all of the [`Filters`].
-struct PaperFiltersIter<'a> {
-    /// The current index of the iteration.
-    index: usize,
-    /// The filters to be iterated.
-    filters: &'a PaperFilters,
-}
-
-impl<'a> Iterator for PaperFiltersIter<'a> {
-    type Item = &'a dyn Filter;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.index += 1;
-
-        match self.index {
-            1 => Some(&self.filters.line),
-            2 => Some(&self.filters.pattern),
-            _ => None,
-        }
-    }
-}
-
-/// Signfifies the view of the current file.
-#[derive(Clone, Debug, Default)]
-struct View {
-    path: PathBuf,
-    /// The data.
-    data: String,
-    /// The first line that is displayed in the ui.
-    first_line: LineNumber,
-    /// The number of columns needed to display the margin.
-    margin_width: usize,
-    /// The number of lines in the data.
-    line_count: usize,
-}
-
-impl View {
-    fn change(&mut self, explorer: Rc<RefCell<dyn Explorer>>, path: PathBuf) -> Outcome<()> {
-        self.data = explorer.borrow_mut().read(&path)?;
-        self.path = path;
-        self.clean();
-        Ok(())
-    }
-
-    /// Adds a character at a [`Mark`].
-    fn add(&mut self, mark: &Mark, c: char) -> Result<(), TryFromIntError> {
-        if let Some(index) = mark.pointer.0 {
-            let data_index = usize::try_from(index)?;
-
-            if let BACKSPACE = c {
-                // For now, do not care to check what is removed. But this may become important for
-                // multi-byte characters.
-                match self.data.remove(data_index) {
-                    _ => {}
-                }
-            } else {
-                self.data.insert(data_index.saturating_sub(1), c);
+            for edit in edits {
+                self.ui.apply(edit)?;
             }
         }
 
         Ok(())
     }
 
-    /// Iterates through the indexes that indicate where each line starts.
-    fn line_indices(&self) -> impl Iterator<Item = IndexType> + '_ {
-        iter::once(0).chain(self.data.match_indices(ENTER).flat_map(|(index, _)| {
-            index
-                .checked_add(1)
-                .and_then(|i| IndexType::try_from(i).ok())
-                .into_iter()
-        }))
-    }
-
-    /// Returns the first column at which view data can be written.
-    #[allow(clippy::integer_arithmetic)] // self.margin_width < usize.max_value()
-    fn first_data_column(&self) -> Result<Index, TryFromIntError> {
-        Index::try_from(self.margin_width + 1)
-    }
-
-    /// Returns the [`Address`] associated with the given [`Place`].
-    fn address_at(&self, place: Place) -> Option<Address> {
-        match Index::try_from(place.line - self.first_line) {
-            Ok(row) => self
-                .first_data_column()
-                .ok()
-                .map(|origin| Address::new(row, place.column + origin)),
-            _ => None,
-        }
-    }
-
-    /// Returns the [`Region`] associated with the given [`Area`].
-    fn region_at<T: Area>(&self, area: &T) -> Option<Region> {
-        self.address_at(area.start())
-            .map(|address| Region::new(address, area.length()))
-    }
-
-    /// Updates the ui with the view's current data.
-    fn redraw_edits(&self) -> impl Iterator<Item = Edit> + '_ {
-        // Clear the screen, then add each row.
-        iter::once(Edit::new(Region::default(), Change::Clear)).chain(
-            self.first_line
-                .into_iter()
-                .zip(self.lines().skip(self.first_line.row()))
-                .flat_map(move |(line_number, line)| {
-                    self.region_at(&Section::line(line_number))
-                        .map(|region| {
-                            Edit::new(
-                                region,
-                                Change::Row(format!(
-                                    "{:>width$} {}",
-                                    line_number,
-                                    line,
-                                    width = self.margin_width
-                                )),
-                            )
-                        })
-                        .into_iter()
-                }),
-        )
-    }
-
-    /// An [`Iterator`] of all lines in the view's data.
-    fn lines(&self) -> std::str::Lines<'_> {
-        self.data.lines()
-    }
-
-    /// The data stored at the given [`LineNumber`].
-    fn line(&self, line_number: LineNumber) -> Option<&str> {
-        self.lines().nth(line_number.row())
-    }
-
-    /// Updates the view's metadata.
-    fn clean(&mut self) {
-        self.line_count = self.lines().count();
-        self.update_margin_width()
-    }
-
-    /// Updates the margin width of view.
-    #[allow(clippy::cast_possible_truncation)] // usize.log10().ceil() < usize.max_value()
-    #[allow(clippy::cast_precision_loss)] // self.line_count is small enough to be precisely represented by f64
-    #[allow(clippy::cast_sign_loss)] // self.line_count >= 0, thus log10().ceil() >= 0.0
-    fn update_margin_width(&mut self) {
-        self.margin_width = (((self.line_count.saturating_add(1)) as f64).log10().ceil()) as usize;
-    }
-
-    /// Scrolls the view's data.
-    fn scroll(&mut self, movement: IndexType) -> IsChanging {
-        let new_first_line = cmp::min(
-            self.first_line + movement,
-            LineNumber::new(self.line_count).unwrap_or_default(),
-        );
-
-        if new_first_line == self.first_line {
-            false
-        } else {
-            self.first_line = new_first_line;
-            true
-        }
-    }
-
-    /// The length of the line that has a given [`Place`].
-    fn line_length(&self, place: Place) -> Option<Index> {
-        self.line(place.line)
-            .and_then(|x| Index::try_from(x.len()).ok())
-    }
-}
-
-impl Hash for View {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.data.hash(state);
-        self.first_line.hash(state);
-        self.margin_width.hash(state);
-        self.line_count.hash(state);
-    }
-}
-
-impl PartialEq for View {
-    fn eq(&self, other: &View) -> bool {
-        self.data == other.data && self.first_line == other.first_line && self.margin_width == other.margin_width && self.line_count == other.line_count
-    }
-}
-
-impl Eq for View {}
-
-/// Represents if the user interface display needs to change.
-type IsChanging = bool;
-
-/// Signifies a modification of the view.
-#[derive(Clone, Debug, Default)]
-struct Adjustment {
-    /// The change made to the current line.
-    shift: IndexType,
-    /// The changes made to the number of lines.
-    line_change: IndexType,
-    /// A map of the indexes where a change was made.
-    indexes_changed: HashMap<LineNumber, IndexType>,
-    /// The [`Change`] that best represents the `Adjustment`.
-    change: Change,
-}
-
-impl Adjustment {
-    /// Creates a new `Adjustment`.
-    fn new(line: LineNumber, shift: IndexType, index_change: IndexType, change: Change) -> Self {
-        let line_change = if change == Change::Clear { shift } else { 0 };
-
-        Self {
-            shift,
-            line_change,
-            indexes_changed: [(line + line_change, index_change)]
-                .iter()
-                .cloned()
-                .collect(),
-            change,
-        }
-    }
-
-    /// Creates an `Adjustment` based on the given context.
-    fn create(c: char, place: Place, view: &View) -> Option<Self> {
-        match c {
-            BACKSPACE => {
-                if place.column == 0 {
-                    view.line_length(place).map(|x| {
-                        Self::new(place.line, NEGATIVE_ONE, IndexType::from(x), Change::Clear)
-                    })
-                } else {
-                    Some(Self::new(
-                        place.line,
-                        NEGATIVE_ONE,
-                        NEGATIVE_ONE,
-                        Change::Backspace,
-                    ))
-                }
-            }
-            ENTER => Some(Self::new(
-                place.line,
-                1,
-                place.column.negate(),
-                Change::Clear,
-            )),
-            _ => Some(Self::new(place.line, 1, 1, Change::Insert(c))),
-        }
-    }
-}
-
-impl AddAssign for Adjustment {
-    fn add_assign(&mut self, other: Self) {
-        self.shift += other.shift;
-        self.line_change += other.line_change;
-
-        for (line, change) in other.indexes_changed {
-            *self.indexes_changed.entry(line).or_default() += change;
-        }
-
-        if self.change != Change::Clear {
-            self.change = other.change
-        }
-    }
-}
-
-/// Indicates a specific Place of a given Section.
-#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
-pub enum Edge {
-    /// Indicates the first Place of the Section.
-    Start,
-    /// Indicates the last Place of the Section.
-    End,
-}
-
-impl Default for Edge {
-    #[inline]
-    fn default() -> Self {
-        Edge::Start
-    }
-}
-
-impl Display for Edge {
-    #[inline]
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match self {
-            Edge::Start => write!(f, "Starting edge"),
-            Edge::End => write!(f, "Ending edge"),
-        }
-    }
-}
-
-/// An address and its respective pointer in a view.
-#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default)]
-struct Mark {
-    /// Pointer in view that corresponds with mark.
-    pointer: Pointer,
-    /// Place of mark.
-    place: Place,
-}
-
-impl Mark {
-    /// Moves `Mark` as specified by the given [`Adjustment`].
-    fn adjust(&mut self, adjustment: &Adjustment) {
-        self.pointer += adjustment.shift;
-        self.place.line = self.place.line + adjustment.line_change;
-
-        for (&line, &change) in &adjustment.indexes_changed {
-            if line == self.place.line {
-                self.place >>= change;
-            }
-        }
-    }
-}
-
-impl Display for Mark {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "{}{}", self.place, self.pointer)
-    }
-}
-
-/// Signifies an index of a character within [`View`].
-#[derive(Copy, Clone, Eq, PartialEq, PartialOrd, Ord, Hash, Debug)]
-struct Pointer(Option<Index>);
-
-impl PartialEq<IndexType> for Pointer {
-    fn eq(&self, other: &IndexType) -> bool {
-        self.0.map_or(false, |x| x == *other)
-    }
-}
-
-impl PartialOrd<IndexType> for Pointer {
-    fn partial_cmp(&self, other: &IndexType) -> Option<Ordering> {
-        self.0.and_then(|x| x.partial_cmp(other))
-    }
-}
-
-impl<T: Borrow<IndexType>> Add<T> for Pointer {
-    type Output = Self;
-
-    fn add(self, other: T) -> Self::Output {
-        Self(self.0.map(|x| x + *other.borrow()))
-    }
-}
-
-impl<T: Borrow<IndexType>> AddAssign<T> for Pointer {
-    fn add_assign(&mut self, other: T) {
-        self.0 = self.0.map(|x| x + *other.borrow());
-    }
-}
-
-impl Default for Pointer {
-    fn default() -> Self {
-        Self(Some(Index::from(0)))
-    }
-}
-
-impl Display for Pointer {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "[{}]",
-            match self.0 {
-                None => String::from("None"),
-                Some(i) => format!("{}", i),
-            }
-        )
-    }
-}
-
-impl PartialEq<Pointer> for IndexType {
-    #[inline]
-    fn eq(&self, other: &Pointer) -> bool {
-        other == self
-    }
-}
-
-impl PartialOrd<Pointer> for IndexType {
-    #[inline]
-    fn partial_cmp(&self, other: &Pointer) -> Option<Ordering> {
-        other.partial_cmp(self).map(|x| x.reverse())
-    }
-}
-
-/// Signifies a type that can be converted to a [`Region`].
-trait Area {
-    /// Returns the starting `Place`.
-    fn start(&self) -> Place;
-    /// Returns the [`Length`].
-    fn length(&self) -> Length;
-}
-
-/// Signifies adjacent [`Place`]s.
-#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug, Default)]
-pub struct Section {
-    /// The [`Place`] at which `Section` starts.
-    start: Place,
-    /// The [`Length`] of `Section`.
-    length: Length,
-}
-
-impl Section {
-    /// Creates a new `Section` that signifies an entire line.
-    #[inline]
-    fn line(line: LineNumber) -> Self {
-        Self {
-            start: Place {
-                line,
-                column: Index::from(0),
-            },
-            length: Length::End,
-        }
-    }
-}
-
-impl Area for Section {
-    fn start(&self) -> Place {
-        self.start
-    }
-
-    fn length(&self) -> Length {
-        self.length
-    }
-}
-
-impl Display for Section {
-    #[inline]
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "{}->{}", self.start, self.length)
-    }
-}
-
-/// Signifies the location of a character within a view.
-#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default)]
-pub struct Place {
-    /// The [`LineNumber`] of `Place`.
-    line: LineNumber,
-    /// The [`Index`] of the column of `Place`.
-    column: Index,
-}
-
-impl Area for Place {
-    fn start(&self) -> Place {
-        *self
-    }
-
-    fn length(&self) -> Length {
-        Length::from(1)
-    }
-}
-
-impl Shr<IndexType> for Place {
-    type Output = Self;
-
-    #[inline]
-    fn shr(self, rhs: IndexType) -> Self {
-        let mut new_place = self;
-        new_place >>= rhs;
-        new_place
-    }
-}
-
-impl ShrAssign<IndexType> for Place {
-    #[inline]
-    fn shr_assign(&mut self, rhs: IndexType) {
-        self.column += rhs;
-    }
-}
-
-impl Display for Place {
-    #[inline]
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "ln {}, idx {}", self.line, self.column)
-    }
-}
-
-/// The type of the value stored in [`LineNumber`].
-type LineNumberType = u32;
-
-/// Signifies a line number.
-#[derive(Copy, Clone, PartialEq, Eq, Ord, PartialOrd, Hash, Debug)]
-struct LineNumber(LineNumberType);
-
-impl LineNumber {
-    /// Creates a new `LineNumber`.
-    fn new(value: usize) -> Option<Self> {
-        if value == 0 {
-            None
-        } else {
-            LineNumberType::try_from(value).ok().map(Self)
-        }
-    }
-
-    /// Converts `LineNumber` to its row index - assuming line number `1` as at row `0`.
-    #[allow(clippy::integer_arithmetic)] // self.0 > 0
-    fn row(self) -> usize {
-        (self.0 - 1) as usize
-    }
-}
-
-impl Add<IndexType> for LineNumber {
-    type Output = Self;
-
-    fn add(self, other: IndexType) -> Self::Output {
-        #[allow(clippy::integer_arithmetic)] // i64::min_value() <= u32 + i32 <= i64::max_value()
-        match usize::try_from(i64::from(self.0) + i64::from(other)) {
-            Ok(sum) => Self::new(sum).unwrap_or_default(),
-            Err(TryFromIntError::Underflow) => Self::default(),
-            Err(TryFromIntError::Overflow) => Self(LineNumberType::max_value()),
-        }
-    }
-}
-
-impl Sub for LineNumber {
-    type Output = i64;
-
-    #[allow(clippy::integer_arithmetic)] // self.0 and other.0 <= u32::MAX
-    fn sub(self, other: Self) -> Self::Output {
-        i64::from(self.0) - i64::from(other.0)
-    }
-}
-
-impl Display for LineNumber {
-    #[inline]
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl Default for LineNumber {
-    #[inline]
-    fn default() -> Self {
-        Self(1)
-    }
-}
-
-impl std::str::FromStr for LineNumber {
-    type Err = ParseLineNumberError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Self::new(s.parse::<usize>()?).ok_or(ParseLineNumberError::InvalidValue)
-    }
-}
-
-impl IntoIterator for LineNumber {
-    type Item = Self;
-    type IntoIter = LineNumberIterator;
-
-    fn into_iter(self) -> Self::IntoIter {
-        LineNumberIterator { current: self }
-    }
-}
-
-/// Signifies an [`Iterator`] of [`LineNumber`]s that steps by 1.
-struct LineNumberIterator {
-    /// The current [`LineNumber`].
-    current: LineNumber,
-}
-
-impl Iterator for LineNumberIterator {
-    type Item = LineNumber;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let line_number = LineNumber(self.current.0);
-        self.current.0 += 1;
-        Some(line_number)
-    }
-}
-
-/// Signifies an error that occurs while parsing a [`LineNumber`] from a [`String`].
-#[derive(Debug)]
-enum ParseLineNumberError {
-    /// The parsed number was not a valid line number.
-    InvalidValue,
-    /// There was an issue parsing the given string to an integer.
-    ParseInt(std::num::ParseIntError),
-}
-
-impl std::error::Error for ParseLineNumberError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match *self {
-            ParseLineNumberError::InvalidValue => None,
-            ParseLineNumberError::ParseInt(ref err) => Some(err),
-        }
-    }
-}
-
-impl Display for ParseLineNumberError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match *self {
-            ParseLineNumberError::InvalidValue => write!(f, "Invalid line number provided."),
-            ParseLineNumberError::ParseInt(ref err) => write!(f, "{}", err),
-        }
-    }
-}
-
-impl From<std::num::ParseIntError> for ParseLineNumberError {
-    fn from(error: std::num::ParseIntError) -> Self {
-        ParseLineNumberError::ParseInt(error)
-    }
-}
-
-/// Used for modifying [`Section`]s to match a feature.
-trait Filter: Debug {
-    /// Returns the identifying character of the `Filter`.
-    fn id(&self) -> char;
-    /// Modifies `sections` such that it matches the given feature.
-    fn extract(
-        &self,
-        feature: &str,
-        sections: &mut Vec<Section>,
-        view: &View,
-    ) -> Result<(), TryFromIntError>;
-}
-
-/// The [`Filter`] used to match a line.
-#[derive(Debug)]
-struct LineFilter {
-    /// The [`Pattern`] used to match one or more [`LineNumber`]s.
-    pattern: Pattern,
-}
-
-impl Default for LineFilter {
-    fn default() -> Self {
-        Self {
-            pattern: Pattern::define(
-                "#" + ((tkn!(some(Digit) => "line") + End)
-                    | (tkn!(some(Digit) => "start") + "." + tkn!(some(Digit) => "end"))
-                    | (tkn!(some(Digit) => "origin") + tkn!(Sign + some(Digit) => "movement"))),
-            ),
-        }
-    }
-}
-
-impl Filter for LineFilter {
-    fn id(&self) -> char {
-        '#'
-    }
-
-    fn extract(
-        &self,
-        feature: &str,
-        sections: &mut Vec<Section>,
-        _view: &View,
-    ) -> Result<(), TryFromIntError> {
-        let tokens = self.pattern.tokenize(feature);
-
-        if let Ok(line) = tokens.parse::<LineNumber>("line") {
-            sections.retain(|&x| x.start.line == line);
-        } else if let (Ok(start), Ok(end)) = (
-            tokens.parse::<LineNumber>("start"),
-            tokens.parse::<LineNumber>("end"),
-        ) {
-            let top = cmp::min(start, end);
-            let bottom = cmp::max(start, end);
-
-            sections.retain(|&x| {
-                let row = x.start.line;
-                row >= top && row <= bottom
-            })
-        } else if let (Ok(origin), Ok(movement)) = (
-            tokens.parse::<LineNumber>("origin"),
-            tokens.parse::<IndexType>("movement"),
-        ) {
-            let end = origin + movement;
-            let top = cmp::min(origin, end);
-            let bottom = cmp::max(origin, end);
-
-            sections.retain(|&x| {
-                let row = x.start.line;
-                row >= top && row <= bottom
-            })
-        }
-
-        Ok(())
-    }
-}
-
-/// A [`Filter`] that extracts matches of a [`Pattern`].
-#[derive(Debug)]
-struct PatternFilter {
-    /// The [`Pattern`] used to match patterns.
-    pattern: Pattern,
-}
-
-impl Default for PatternFilter {
-    fn default() -> Self {
-        Self {
-            pattern: Pattern::define("/" + tkn!(some(Any) => "pattern")),
-        }
-    }
-}
-
-impl Filter for PatternFilter {
-    fn id(&self) -> char {
-        '/'
-    }
-
-    fn extract(
-        &self,
-        feature: &str,
-        sections: &mut Vec<Section>,
-        view: &View,
-    ) -> Result<(), TryFromIntError> {
-        if let Some(user_pattern) = self.pattern.tokenize(feature).get("pattern") {
-            if let Ok(search_pattern) = Pattern::load(user_pattern) {
-                let target_sections = sections.clone();
-                sections.clear();
-
-                for target_section in target_sections {
-                    let start = usize::try_from(target_section.start.column)?;
-
-                    if let Some(target) = view
-                        .line(target_section.start.line)
-                        .map(|x| x.chars().skip(start).collect::<String>())
-                    {
-                        for location in search_pattern.locate_iter(&target) {
-                            sections.push(Section {
-                                start: target_section.start
-                                    >> IndexType::try_from(location.start())?,
-                                length: Length::try_from(location.length())?,
-                            });
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(())
+    fn current_processor_mut(&mut self) -> &mut Rc<RefCell<dyn Processor>> {
+        self.mode_handlers.get_mut(&self.mode).unwrap()
     }
 }
